level 2-7 원 넓이 계산

+++ src/main/java/calculator/Calculator.java
@@ -6,9 +6,11 @@ import java.util.List;
 public class Calculator {
 
     private final List<Integer> db;
+    private final List<Double> circleDb;
 
     public Calculator() {
         db = new ArrayList<>();
+        circleDb = new ArrayList<>();
     }
 
     public int calculate(int a, int b, char op) {
@@ -34,19 +36,39 @@ public class Calculator {
         }
     }
 
-    public void save(int result) {
+    public double calculateCircleArea(int radius) {
+        return Math.PI * radius * radius; // 항성 고정된 정적 변수 값 Math.PI
+    }
+
+    public void saveResult(int result) {
         db.add(result);
     }
 
-    public void removeFirst() {
-        db.remove(0);
+    public void saveCircleAreaResult(double result) {
+        circleDb.add(result);
+    }
+
+    public void removeFirst(String type) {
+        if (type.equals("normal")) {
+            db.remove(0);
+        }
+
+        if (type.equals("circle")) {
+            circleDb.remove(0);
+        }
     }
 
     public void printList() {
+        System.out.print("일반 계산 목록 >> ");
         for (int i : db) {
             System.out.print(i + " ");
         }
+        System.out.println();
 
+        System.out.print("원 넓이 목록 >> ");
+        for (double r : circleDb) {
+            System.out.print(r + " ");
+        }
         System.out.println();
     }
 }
